files/source/requirements.txt
kafka-python
polars
boto3
s3fs
pyarrow
faker  # Para generar datos falsos realistas